chore: separate types and update usage
