+++ src/hooks/useTransactionHistory.tsx
@@ -1,5 +1,5 @@
 import { useInfiniteQuery, UseInfiniteQueryResult } from '@tanstack/react-query';
-import { FilterOptions } from 'src/modules/history/HistoryFilterMenu';
+import { actionFilterMap, HistoryFilters, TransactionHistoryItem } from 'src/modules/history/types';
 import {
   USER_TRANSACTIONS_V2,
   USER_TRANSACTIONS_V2_WITH_POOL,
@@ -7,109 +7,6 @@ import {
 import { USER_TRANSACTIONS_V3 } from 'src/modules/history/v3-user-history-query';
 import { useRootStore } from 'src/store/root';
 import { QueryKeys } from 'src/ui-config/queries';
-
-export type TransactionHistoryItem<T = unknown> = {
-  id: string;
-  txHash: string;
-  action: string;
-  pool: {
-    id: string;
-  };
-  timestamp: number;
-} & T;
-
-type ReserveSubset = {
-  symbol: string;
-  decimals: number;
-  underlyingAsset: string;
-  name: string;
-};
-
-export type ActionFields = {
-  Supply: {
-    reserve: ReserveSubset;
-    amount: string;
-    assetPriceUSD: string;
-  };
-  Deposit: {
-    reserve: ReserveSubset;
-    amount: string;
-    assetPriceUSD: string;
-  };
-  Borrow: {
-    reserve: ReserveSubset;
-    amount: string;
-    assetPriceUSD: string;
-  };
-  Repay: {
-    reserve: ReserveSubset;
-    amount: string;
-    assetPriceUSD: string;
-  };
-  RedeemUnderlying: {
-    reserve: ReserveSubset;
-    amount: string;
-    assetPriceUSD: string;
-  };
-  LiquidationCall: {
-    collateralReserve: ReserveSubset;
-    collateralAmount: string;
-    principalReserve: ReserveSubset;
-    principalAmount: string;
-    borrowAssetPriceUSD: string;
-    collateralAssetPriceUSD: string;
-  };
-  SwapBorrowRate: {
-    reserve: ReserveSubset;
-    borrowRateModeFrom: string;
-    borrowRateModeTo: string;
-    stableBorrowRate: string;
-    variableBorrowRate: string;
-    assetPriceUSD: string;
-  };
-  Swap: {
-    reserve: ReserveSubset;
-    borrowRateModeFrom: number;
-    borrowRateModeTo: number;
-    stableBorrowRate: string;
-    variableBorrowRate: string;
-    assetPriceUSD: string;
-  };
-  UsageAsCollateral: {
-    reserve: ReserveSubset;
-    fromState: boolean;
-    toState: boolean;
-    assetPriceUSD: string;
-  };
-};
-
-interface HistoryFilters {
-  searchQuery: string;
-  filterQuery: FilterOptions[];
-}
-
-const actionFilterMap = (action: string): number => {
-  switch (action) {
-    case 'Deposit': // v2
-    case 'Supply': // v3
-      return 0;
-    case 'Borrow':
-      return 1;
-    case 'RedeemUnderlying':
-      return 2;
-    case 'Repay':
-      return 3;
-    case 'Swap': // v2
-    case 'SwapBorrowRate': // v3
-      return 4;
-    case 'UsageAsCollateral':
-      return 5;
-    case 'LiquidationCall':
-      return 6;
-    default:
-      return 7;
-  }
-};
 
 export const applyTxHistoryFilters = ({
   searchQuery,
@@ -120,7 +17,7 @@ export const applyTxHistoryFilters = ({
 
   // Apply seach filter
   if (searchQuery.length > 0) {
-    // txn may or may not contain reserve fields, there's definitely a way to handle this with generics but I'm lazy
+    // txn may or may not contain certain reserve fields depending on tx type
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     filteredTxns = txns.filter((txn: any) => {
       const collateralSymbol = txn?.collateralReserve?.symbol?.toLowerCase(); // for liquidationcall
